+++ libs/api/domains/work-machines/src/lib/dto/getDocuments.input.ts
@@ -1,5 +1,5 @@
 import { Field, InputType, registerEnumType } from '@nestjs/graphql'
-import { FileType } from '../api-domains-work-machines.types'
+import { FileType } from '../workMachines.types'
 import { IsEnum, IsOptional } from 'class-validator'
 
 registerEnumType(FileType, { name: 'WorkMachinesFileType' })
