+++ libs/api/domains/work-machines/src/lib/workMachines.types.ts
